[refactor/#30] - 출석 관련 도메인 모델 및 Enum 파일 분리
