+++ domain/src/main/java/com/umc/domain/model/enums/CheckAvailableStatus.kt
@@ -1,0 +1,7 @@
+package com.umc.domain.model.enums
+
+enum class CheckAvailableStatus(val text: String) {
+    BEFORE("출석 전"),
+    PENDING("승인 대기"),
+    COMPLETED("출석 완료")
+}
